removed auxillary files
